+++ src/components/myPage/index.js
@@ -2,7 +2,8 @@ export { default as MypageTitle } from './MypageTitle';
 export { default as CollectionButton } from './CollectionButton';
 export { default as Collection } from './Collection';
 export { default as Collections } from './Collections';
-export { default as SubscribeBadges } from './SubscribeBadges';
+export { default as ProviderBadges } from './ProviderBadges';
 export { default as Statistics } from './Statistics';
-export { default as SubscriptionServices } from './SubscriptionServices';
+export { default as SubscriptionProviders } from './SubscriptionProviders';
 export { default as CurrentSubscriptionInfo } from './CurrentSubscriptionInfo';
+export { default as ProviderChips } from './ProviderChips';
